// src/app/types/ResTypes.ts
export interface SignUpRes {
  status: boolean;
  data: {
    token: string;
    user: {
      user_name: string;
      email: string;
      user_id:number;
    };
    newUser: boolean;
  };
  message: string;
  toast: boolean;
}